+++ src/error.cpp
@@ -11,6 +11,8 @@ error::error(error_code code, std::string message)
 {
 }
 
+error::~error() = default;
+
 error_code error::code() const noexcept { return code_; }
 
 } // namespace pkgexec_linux